Fix return type for convert_list_to_bitmap()

+++ src/c/pyeclib_c/pyeclib_c.c
@@ -151,10 +151,10 @@ static int validate_args(int k, int m, int w, pyeclib_type_t type)
  * Convert an int list into a bitmap
  * Assume the list is '-1' terminated.
  */
-static unsigned long convert_list_to_bitmap(int *list)
+static unsigned long long convert_list_to_bitmap(int *list)
 {
   int i = 0;
-  unsigned long bm = 0;
+  unsigned long long bm = 0;
 
   while (list[i] > -1) {
     /*
